requirements.txt
contourpy==1.1.0
cycler==0.11.0
fonttools==4.40.0
importlib-resources==6.0.0
kiwisolver==1.4.4
matplotlib==3.7.2
numpy==1.24.4
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
six==1.16.0
tqdm==4.65.0
tzdata==2023.3
zipp==3.15.0
